requirements.txt
appdirs==1.4.3
Django==2.2.1
djangorestframework==3.9.4
packaging==19.0
pyparsing==2.4.0
pytz==2019.1
six==1.12.0
uWSGI==2.0.18
psycopg2==2.8.2